requirements.txt
Django==2.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-guardian==1.4.9
django-markdown==0.8.4
django-markdown-deux==1.0.5
Markdown==2.6.11
markdown2==2.3.5
Pillow==5.2.0
pytz==2018.5
six==1.11.0
djangorestframework==3.8.2
